[DEV-5735] Move and prettify calc_metrics_jira_linear
