+++ server/tests/controllers/test_jira_controller.py
@@ -1,9 +1,7 @@
-from datetime import date, datetime, timedelta, timezone
+from datetime import datetime, timedelta, timezone
 import json
-from typing import List
 
 from dateutil.tz import tzutc
-import numpy as np
 import pytest
 from sqlalchemy import delete, insert, update
 
@@ -12,11 +10,8 @@ from athenian.api.internal.miners.filters import JIRAFilter, LabelFilter
 from athenian.api.internal.settings import LogicalRepositorySettings
 from athenian.api.models.metadata.github import NodePullRequestJiraIssues, PullRequest
 from athenian.api.models.metadata.jira import Issue
-from athenian.api.models.state.models import MappedJIRAIdentity
 from athenian.api.models.web import (
     CalculatedJIRAHistogram,
-    CalculatedJIRAMetricValues,
-    CalculatedLinearMetricValues,
     DeployedComponent,
     DeploymentNotification,
     FilteredJIRAStuff,
@@ -1450,556 +1445,6 @@ async def test_filter_jira_nasty_input(client, headers, account, date_to, tz, st
     )
     body = (await response.read()).decode("utf-8")
     assert response.status == status, "Response body is : " + body
-
-
-@pytest.mark.parametrize("exclude_inactive", [False, True])
-async def test_jira_metrics_smoke(client, headers, exclude_inactive):
-    body = {
-        "date_from": "2020-01-01",
-        "date_to": "2020-10-23",
-        "timezone": 120,
-        "account": 1,
-        "metrics": [JIRAMetricID.JIRA_RAISED, JIRAMetricID.JIRA_RESOLVED],
-        "exclude_inactive": exclude_inactive,
-        "granularities": ["all", "2 month"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    body = json.loads(body)
-    assert len(body) == 2
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in body]
-    assert items[0].granularity == "all"
-    assert items[0].with_ is None
-    assert items[0].values == [
-        CalculatedLinearMetricValues(
-            date=date(2020, 1, 1),
-            values=[1699, 1573],
-            confidence_mins=[None] * 2,
-            confidence_maxs=[None] * 2,
-            confidence_scores=[None] * 2,
-        ),
-    ]
-    assert items[1].granularity == "2 month"
-    assert items[1].with_ is None
-    assert len(items[1].values) == 5
-    assert items[1].values[0] == CalculatedLinearMetricValues(
-        date=date(2020, 1, 1),
-        values=[160, 39],
-        confidence_mins=[None] * 2,
-        confidence_maxs=[None] * 2,
-        confidence_scores=[None] * 2,
-    )
-    assert items[1].values[-1] == CalculatedLinearMetricValues(
-        date=date(2020, 9, 1),
-        values=[237, 238],
-        confidence_mins=[None] * 2,
-        confidence_maxs=[None] * 2,
-        confidence_scores=[None] * 2,
-    )
-
-
-@pytest.mark.parametrize(
-    "account, metrics, date_from, date_to, timezone, granularities, status",
-    [
-        (1, [JIRAMetricID.JIRA_RAISED], "2020-01-01", "2020-04-01", 120, ["all"], 200),
-        (2, [JIRAMetricID.JIRA_RAISED], "2020-01-01", "2020-04-01", 120, ["all"], 422),
-        (3, [JIRAMetricID.JIRA_RAISED], "2020-01-01", "2020-04-01", 120, ["all"], 404),
-        (1, [], "2020-01-01", "2020-04-01", 120, ["all"], 200),
-        (1, None, "2020-01-01", "2020-04-01", 120, ["all"], 400),
-        (1, [JIRAMetricID.JIRA_RAISED], "2020-05-01", "2020-04-01", 120, ["all"], 400),
-        (1, [JIRAMetricID.JIRA_RAISED], "2020-01-01", "2020-04-01", 100500, ["all"], 400),
-        (1, [JIRAMetricID.JIRA_RAISED], "2020-01-01", "2020-04-01", 120, ["whatever"], 400),
-    ],
-)
-async def test_jira_metrics_nasty_input1(
-    client,
-    headers,
-    account,
-    metrics,
-    date_from,
-    date_to,
-    timezone,
-    granularities,
-    status,
-):
-    body = {
-        "date_from": date_from,
-        "date_to": date_to,
-        "timezone": timezone,
-        "account": account,
-        "metrics": metrics,
-        "with": [],
-        "exclude_inactive": True,
-        "granularities": granularities,
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == status, "Response body is : " + body
-
-
-async def test_jira_metrics_priorities(client, headers):
-    body = {
-        "date_from": "2020-01-01",
-        "date_to": "2020-10-20",
-        "timezone": 0,
-        "account": 1,
-        "metrics": [JIRAMetricID.JIRA_RAISED],
-        "exclude_inactive": True,
-        "granularities": ["all"],
-        "priorities": ["high"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    body = json.loads(body)
-    assert len(body) == 1
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in body]
-    assert items[0].granularity == "all"
-    assert items[0].values[0].values == [392]
-
-
-async def test_jira_metrics_types(client, headers):
-    body = {
-        "date_from": "2020-01-01",
-        "date_to": "2020-10-20",
-        "timezone": 0,
-        "account": 1,
-        "metrics": [JIRAMetricID.JIRA_RAISED],
-        "exclude_inactive": True,
-        "granularities": ["all"],
-        "types": ["tASK"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    body = json.loads(body)
-    assert len(body) == 1
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in body]
-    assert items[0].granularity == "all"
-    assert items[0].values[0].values == [658]
-
-
-async def test_jira_metrics_epics(client, headers):
-    body = {
-        "date_from": "2020-01-01",
-        "date_to": "2020-10-20",
-        "timezone": 0,
-        "account": 1,
-        "metrics": [JIRAMetricID.JIRA_RAISED],
-        "exclude_inactive": True,
-        "granularities": ["all"],
-        "epics": ["DEV-70", "DEV-843"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    body = json.loads(body)
-    assert len(body) == 1
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in body]
-    assert items[0].granularity == "all"
-    assert items[0].values[0].values == [38]
-
-
-async def test_jira_metrics_labels(client, headers):
-    body = {
-        "date_from": "2020-01-01",
-        "date_to": "2020-10-20",
-        "timezone": 0,
-        "account": 1,
-        "metrics": [JIRAMetricID.JIRA_RAISED],
-        "exclude_inactive": True,
-        "granularities": ["all"],
-        "labels_include": ["PERFORmance"],
-        "labels_exclude": ["buG"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    body = json.loads(body)
-    assert len(body) == 1
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in body]
-    assert items[0].granularity == "all"
-    # 148 without labels_exclude
-    # 147 without cleaning
-    assert items[0].values[0].values == [142]
-
-
-@pytest.mark.parametrize(
-    "assignees, reporters, commenters, count",
-    [
-        (["Vadim markovtsev"], ["waren long"], ["lou Marvin caraig"], 1136),
-        (["Vadim markovtsev"], [], [], 529),
-        ([], [], ["{1}"], 236),
-        ([None, "Vadim MARKOVTSEV"], [], [], 694),
-        ([], ["waren long"], [], 539),
-        ([], [], ["lou Marvin caraig"], 236),
-        ([None], [], ["lou Marvin caraig"], 381),
-    ],
-)
-async def test_jira_metrics_people(
-    client,
-    headers,
-    assignees,
-    reporters,
-    commenters,
-    count,
-    sample_team,
-    sdb,
-):
-    await sdb.execute(
-        insert(MappedJIRAIdentity).values(
-            {
-                MappedJIRAIdentity.account_id: 1,
-                MappedJIRAIdentity.confidence: 1,
-                MappedJIRAIdentity.github_user_id: 51,
-                MappedJIRAIdentity.jira_user_id: "5ddec0b9be6c1f0d071ff82d",
-                MappedJIRAIdentity.created_at: datetime.now(timezone.utc),
-                MappedJIRAIdentity.updated_at: datetime.now(timezone.utc),
-            },
-        ),
-    )
-    body = {
-        "date_from": "2020-01-01",
-        "date_to": "2020-10-20",
-        "timezone": 0,
-        "account": 1,
-        "metrics": [JIRAMetricID.JIRA_RAISED],
-        "exclude_inactive": True,
-        "granularities": ["all"],
-        "with": [
-            {
-                "assignees": assignees,
-                "reporters": reporters,
-                "commenters": commenters,
-            },
-        ],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    body = json.loads(body)
-    assert len(body) == 1
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in body]
-    assert items[0].granularity == "all"
-    with_ = {}
-    for key, val in {
-        "assignees": assignees,
-        "reporters": reporters,
-        "commenters": commenters,
-    }.items():
-        if val:
-            with_[key] = val
-    assert items[0].with_.to_dict() == with_
-    assert items[0].values[0].values == [count]
-
-
-async def test_jira_metrics_teams(client, headers):
-    body = {
-        "date_from": "2020-01-01",
-        "date_to": "2020-10-20",
-        "timezone": 0,
-        "account": 1,
-        "metrics": [JIRAMetricID.JIRA_RAISED],
-        "exclude_inactive": True,
-        "granularities": ["all"],
-        "with": [
-            {
-                "assignees": ["vadim Markovtsev"],
-            },
-            {
-                "reporters": ["waren Long"],
-            },
-        ],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    body = json.loads(body)
-    assert len(body) == 2
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in body]
-    assert items[0].granularity == "all"
-    assert items[0].values[0].values == [529]
-    assert items[0].with_.to_dict() == {"assignees": ["vadim Markovtsev"]}
-    assert items[1].values[0].values == [539]
-    assert items[1].with_.to_dict() == {"reporters": ["waren Long"]}
-
-
-@pytest.mark.parametrize(
-    "metric, exclude_inactive, n",
-    [
-        (JIRAMetricID.JIRA_OPEN, False, 197),
-        (JIRAMetricID.JIRA_OPEN, True, 188),
-        (JIRAMetricID.JIRA_RESOLVED, False, 795),
-        (JIRAMetricID.JIRA_RESOLVED, True, 795),
-        (JIRAMetricID.JIRA_ACKNOWLEDGED, False, 776),
-        (JIRAMetricID.JIRA_ACKNOWLEDGED_Q, False, 776),
-        (JIRAMetricID.JIRA_RESOLUTION_RATE, False, 0.9578313231468201),
-    ],
-)
-async def test_jira_metrics_counts(client, headers, metric, exclude_inactive, n):
-    body = {
-        "date_from": "2020-06-01",
-        "date_to": "2020-10-23",
-        "timezone": 120,
-        "account": 1,
-        "metrics": [metric],
-        "exclude_inactive": exclude_inactive,
-        "granularities": ["all"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    body = json.loads(body)
-    assert len(body) == 1
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in body]
-    assert items[0].granularity == "all"
-    assert items[0].values == [
-        CalculatedLinearMetricValues(
-            date=date(2020, 6, 1),
-            values=[n],
-            confidence_mins=[None],
-            confidence_maxs=[None],
-            confidence_scores=[None],
-        ),
-    ]
-
-
-# TODO: fix response validation against the schema
-@pytest.mark.app_validate_responses(False)
-@pytest.mark.parametrize(
-    "metric, value, score, cmin, cmax",
-    [
-        (JIRAMetricID.JIRA_LIFE_TIME, "3360382s", 51, "2557907s", "4227690s"),
-        (JIRAMetricID.JIRA_LEAD_TIME, "2922288s", 43, "2114455s", "3789853s"),
-        (JIRAMetricID.JIRA_ACKNOWLEDGE_TIME, "450250s", 66, "369809s", "527014s"),
-        (JIRAMetricID.JIRA_PR_LAG_TIME, "0s", 100, "0s", "0s"),
-        (JIRAMetricID.JIRA_BACKLOG_TIME, "450856s", 66, "370707s", "527781s"),
-    ],
-)
-@with_defer
-async def test_jira_metrics_bug_times(
-    client,
-    headers,
-    metric,
-    value,
-    score,
-    cmin,
-    cmax,
-    pr_facts_calculator_factory,
-    release_match_setting_tag,
-    prefixer,
-    bots,
-):
-    pr_facts_calculator_no_cache = pr_facts_calculator_factory(1, (6366825,))
-    time_from = datetime(year=2018, month=1, day=1, tzinfo=timezone.utc)
-    time_to = datetime(year=2020, month=4, day=1, tzinfo=timezone.utc)
-    args = (
-        time_from,
-        time_to,
-        {"src-d/go-git"},
-        {},
-        LabelFilter.empty(),
-        JIRAFilter.empty(),
-        False,
-        bots,
-        release_match_setting_tag,
-        LogicalRepositorySettings.empty(),
-        prefixer,
-        False,
-        0,
-    )
-    await pr_facts_calculator_no_cache(*args)
-    await wait_deferred()
-    np.random.seed(7)
-    body = {
-        "date_from": "2016-01-01",
-        "date_to": "2020-10-23",
-        "timezone": 120,
-        "account": 1,
-        "metrics": [metric],
-        "types": ["BUG"],
-        "exclude_inactive": False,
-        "granularities": ["all", "1 year"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    body = json.loads(body)
-    assert len(body) == 2
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in body]
-    assert items[0].granularity == "all"
-    assert items[0].values == [
-        CalculatedLinearMetricValues(
-            date=date(2016, 1, 1),
-            values=[value],
-            confidence_mins=[cmin],
-            confidence_maxs=[cmax],
-            confidence_scores=[score],
-        ),
-    ]
-
-
-async def test_jira_metrics_disabled_projects(client, headers, disabled_dev):
-    body = {
-        "date_from": "2020-01-01",
-        "date_to": "2020-10-23",
-        "timezone": 120,
-        "account": 1,
-        "metrics": [JIRAMetricID.JIRA_RAISED, JIRAMetricID.JIRA_RESOLVED],
-        "exclude_inactive": False,
-        "granularities": ["all", "2 month"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in json.loads(body)]
-    _check_metrics_no_dev_project(items)
-
-
-async def test_jira_metrics_selected_projects(client, headers, client_cache):
-    body = {
-        "date_from": "2020-01-01",
-        "date_to": "2020-10-23",
-        "timezone": 120,
-        "account": 1,
-        "metrics": [JIRAMetricID.JIRA_RAISED, JIRAMetricID.JIRA_RESOLVED],
-        "exclude_inactive": False,
-        "projects": ["PRO", "OPS", "ENG", "GRW", "CS", "CON"],
-        "granularities": ["all", "2 month"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    result = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + result
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in json.loads(result)]
-    _check_metrics_no_dev_project(items)
-
-    del body["projects"]
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    result = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + result
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in json.loads(result)]
-    with pytest.raises(AssertionError):
-        _check_metrics_no_dev_project(items)
-
-
-def _check_metrics_no_dev_project(items: List[CalculatedJIRAMetricValues]) -> None:
-    assert items[0].values == [
-        CalculatedLinearMetricValues(
-            date=date(2020, 1, 1),
-            values=[767, 829],
-            confidence_mins=[None] * 2,
-            confidence_maxs=[None] * 2,
-            confidence_scores=[None] * 2,
-        ),
-    ]
-
-
-async def test_jira_metrics_group_by_label_smoke(client, headers):
-    body = {
-        "date_from": "2020-01-01",
-        "date_to": "2020-10-20",
-        "timezone": 0,
-        "account": 1,
-        "metrics": [JIRAMetricID.JIRA_RAISED],
-        "exclude_inactive": True,
-        "granularities": ["all"],
-        "group_by_jira_label": True,
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    rbody = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + rbody
-    rbody = json.loads(rbody)
-    assert len(rbody) == 49
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in rbody]
-    assert items[0].granularity == "all"
-    assert items[0].jira_label == "performance"
-    assert items[0].values[0].values == [143]
-    assert items[1].jira_label == "webapp"
-    assert items[1].values[0].values == [142]
-    assert items[-1].jira_label is None
-    assert items[-1].values[0].values == [729]
-
-    body["labels_include"] = ["performance"]
-    body["labels_exclude"] = ["security"]
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    rbody = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + rbody
-    rbody = json.loads(rbody)
-    assert len(rbody) == 1
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in rbody]
-    assert items[0].granularity == "all"
-    assert items[0].jira_label == "performance"
-    assert items[0].values[0].values == [142]
-
-
-async def test_jira_metrics_group_by_label_empty(client, headers):
-    body = {
-        "date_from": "2019-12-02",
-        "date_to": "2019-12-03",
-        "timezone": 0,
-        "account": 1,
-        "metrics": [JIRAMetricID.JIRA_RAISED],
-        "exclude_inactive": True,
-        "granularities": ["all"],
-        "group_by_jira_label": True,
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    rbody = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + rbody
-    rbody = json.loads(rbody)
-    assert len(rbody) == 1
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in rbody]
-    assert len(items) == 1
-    assert items[0].granularity == "all"
-    assert items[0].jira_label is None
-    assert items[0].values[0].values == [9]
-
-    body["labels_include"] = ["whatever"]
-    response = await client.request(
-        method="POST", path="/v1/metrics/jira", headers=headers, json=body,
-    )
-    rbody = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + rbody
-    rbody = json.loads(rbody)
-    assert len(rbody) == 1
-    items = [CalculatedJIRAMetricValues.from_dict(i) for i in rbody]
-    assert len(items) == 1
-    assert items[0].granularity == "all"
-    assert items[0].jira_label is None
-    assert items[0].values[0].values == [0]
 
 
 @pytest.mark.parametrize(
